refactor(04-uat): rename shipped→mentioned across changelog mention banner
The "shipped" name was a leftover from when the banner read "Shipped
in <changelog>". After the wording neutralized to "Mentioned in"
(6b0ecb5) the CSS class, partial filename, model scope, locale key,
test class, and stylesheet filename all kept the old name and
drifted out of sync with the user-facing copy.

Rename across the board so the codebase reads consistently:

  partial:   _shipped_banner.html.erb → _mention_banner.html.erb
  stylesheet: shipped-banner.css      → mention-banner.css
  test:       shipped_banner_test.rb  → mention_banner_test.rb
  CSS class:  .shipped-banner         → .mention-banner
  CSS tokens: --shipped-banner-*      → --mention-banner-*
  scope:      idea.shipped_changelogs → idea.mentioned_in_changelogs
  locale:     ideas.shipped_banner.*  → ideas.mention_banner.*
  test class: ShippedBannerTest       → MentionBannerTest

File renames done via `git mv` so history follows. The model scope
keeps its `published` order constraint — it's still "changelogs this
idea is mentioned in that have been released to the public", which
is the correct relationship.

No domain-layer "shipped" semantics changed (idea status names,
release terminology, etc) — the rename is scoped only to the banner
machinery that visually displays the mention relationship.
