--- app/assets/stylesheets/shipped-banner.css
@@ -1,43 +1,0 @@
-/* ==========================================================================
-   Shipped Banner Module
-
-   Renders above the idea title when a published changelog links the idea.
-   Surface color recipe mirrors .alert--success (alerts.css:119-127).
-   ========================================================================== */
-
-@layer modules {
-  .shipped-banner {
-    /* Component API */
-    --shipped-banner-color: var(--color-success);
-    --shipped-banner-border-color: var(--color-success);
-    --shipped-banner-background: oklch(from var(--color-success) l c h / 10%);
-    --shipped-banner-padding-inline: var(--space-4);
-    --shipped-banner-padding-block: var(--space-3);
-    --shipped-banner-radius: var(--radius-md);
-
-    background-color: var(--shipped-banner-background);
-    border: 1px solid var(--shipped-banner-border-color);
-    border-radius: var(--shipped-banner-radius);
-    color: var(--shipped-banner-color);
-    padding-inline: var(--shipped-banner-padding-inline);
-    padding-block: var(--shipped-banner-padding-block);
-  }
-
-  .shipped-banner__link {
-    color: inherit;
-    font-weight: var(--font-semibold);
-    text-decoration: underline;
-    text-underline-offset: 0.15em;
-    text-decoration-thickness: 1px;
-    transition: text-decoration-thickness var(--duration-fast);
-  }
-
-  .shipped-banner__link:hover {
-    text-decoration-thickness: 2px;
-  }
-
-  .shipped-banner__when {
-    margin-inline-start: var(--space-1);
-    opacity: 0.7;
-  }
-}
