--- app/views/ideas/_shipped_banner.html.erb
@@ -1,18 +1,0 @@
-<%# locals: (idea:) %>
-<% shipped = idea.shipped_changelogs.to_a %>
-<% if shipped.any? %>
-  <% changelog = shipped.first %>
-  <aside class="shipped-banner flex align-center gap-2" role="status">
-    <%= lucide_icon "megaphone", class: "size-4 flex-item-no-shrink" %>
-    <p class="txt-small">
-      <%= t(
-        ".mentioned_in_html",
-        changelog: link_to(changelog.title, changelog_path(changelog), class: "shipped-banner__link", data: { turbo_frame: "_top" })
-      ) %>
-      <span class="shipped-banner__when" aria-hidden="true">·</span>
-      <span class="shipped-banner__when">
-        <%= time_ago_with_tooltip(changelog.published_at) %>
-      </span>
-    </p>
-  </aside>
-<% end %>
