+++ app/views/ideas/show.html.erb
@@ -10,7 +10,7 @@
       <% end %>
     </div>
 
-    <%= render "shipped_banner", idea: @idea %>
+    <%= render "mention_banner", idea: @idea %>
 
     <%= render Elements::PanelComponent.new do |c| %>
       <% c.with_body(class: "pad") do %>
